Add support for 64bit hash codes

+++ Src/FastData/Contexts/HashSetLinearContext.cs
@@ -4,12 +4,12 @@ namespace Genbox.FastData.Contexts;
 
 public class HashSetLinearContext<T>(T[] data) : DefaultContext<T>(data)
 {
-    public HashSetLinearContext(T[] data, HashSetBucket[] buckets, uint[] hashCodes) : this(data)
+    public HashSetLinearContext(T[] data, HashSetBucket[] buckets, ulong[] hashCodes) : this(data)
     {
         Buckets = buckets;
         HashCodes = hashCodes;
     }
 
     public HashSetBucket[] Buckets { get; }
-    public uint[] HashCodes { get; }
+    public ulong[] HashCodes { get; }
 }
