+++ Src/FastData/Specs/HashFunc.cs
@@ -1,3 +1,3 @@
 namespace Genbox.FastData.Specs;
 
-public delegate uint HashFunc<in T>(T obj);
+public delegate ulong HashFunc<in T>(T obj);
